Corrigindo erros ao subir o projeto para o GitHub

+++ src/main/java/com/br/sistemapadaria/service/ClienteService.java
@@ -7,6 +7,12 @@ import org.springframework.stereotype.Service;
 import com.br.sistemapadaria.model.Cliente;
 import com.br.sistemapadaria.repository.ClienteRepository;
 
+/*
+ * Descrição: Classe Cliente service
+ * Data: 01/12/2019  22:30
+ * Autor: José Walter
+ */
+
 @Service
 public class ClienteService {
 	
@@ -15,19 +21,19 @@ public class ClienteService {
 		
 		public Cliente cadastrar(Cliente cliente){
 			
-			if(cliente.getCodigoCliente()==0){
-			 clienteRepository.save(cliente);
-			 System.out.println("O cliente " + cliente+ ", foi salvo com sucesso!!");
-		     
-		}else{
-			System.out.println("O cliente já existe!!!");
+			if(cliente != null){
+	    		 clienteRepository.save(cliente);
+		    	 System.out.println("O cliente " + cliente+ ", foi salvo com sucesso!!");
+		    	 	     
+    		}else{
+			    System.out.println("O cliente já existe!!!");
 			
 		}
 			return cliente;
 		}
 		
 		
-	//======================================================================
+	//=======================================================================================
 		// Busca todos os dados do banco
 		public List<Cliente> buscarTodos(){
 			
@@ -35,25 +41,37 @@ public class ClienteService {
 			
 		}
 		
-	//======================================================================
+	//=======================================================================================
 		//Excclui um cliente do banco
 		public void excluir(Cliente cliente){
 			
-			clienteRepository.delete(cliente);
+			if(cliente.getCodigoCliente() != 0){
+				clienteRepository.delete(cliente);
+				System.out.println(" O cliente " + cliente + " foi excluido com sucesso!!");
+			}else{
+				System.out.println(" O cliente não foi encontrado!!");
+			}
+			
 		}
 		
-	//======================================================================
+	//=======================================================================================
 		// Faz a busca de um cliente pelo ID
 		public Cliente buscarPorId(Long id) {
 			
 			return clienteRepository.getOne(id);
 		}
 		
-	//======================================================================
+	//=======================================================================================
 		//Faz a alteração do cliente cadastrado no banco 
 		public Cliente alterar(Cliente cliente){
 			
-			clienteRepository.save(cliente);
+			if(cliente.getCodigoCliente() != 0){
+				clienteRepository.save(cliente);
+				System.out.println(" O cliente " + cliente + " foi alterado com sucesso!!");
+			}else{
+				System.out.println("Não conseguimos fazer a alteração do cliente!!");
+			}
+			
 			return cliente;
 		}
 	}
